revert (partially) 4543b0f
Rationale: the logic before 4543b0f was right. Assuming we use the same
rounding mode before and after mpfr_subnormalize, we get underflow after
rounding iff the input of mpfr_subnormalize is in the subnormal range.

The only difference is that now, the underflow flag is not raised when
the result is an exact result in the subnormal range (i.e., when both
ternary values before and after mpfr_subnormalize are zero).