[tests/thash.c] Code improvement (avoid a warning with -std=c90 -m32).
The issue is the use of a suffixless decimal constant that does not
fit in a 32-bit long. In C90, it is just converted to unsigned long
(the target type), which is fine. But in C99+, it is converted to
long long, then to unsigned long for the assignment; this works here,
but this is a bit ugly and can yield confusion.

So, let's add a U suffix. But shouldn't hexadecimal constants be
better (as these constants really consist of a sequence of bits)?